Replace the data type of `xbt_id`, `cdt_pri_id`, `cdt_awd_pri_id`, `cdt_awd_pri_xps_type_map_id`, `cdt_sc_awd_pri_id`, and `cdt_sc_awd_pri_xps_type_map_id` to char(36) for UUID.

+++ score-http/src/main/java/org/oagi/score/repo/component/bdt_pri_restri/BdtPriRestriReadRepository.java
@@ -38,7 +38,7 @@ public class BdtPriRestriReadRepository {
                 isDefault = "time".equalsIgnoreCase(xbtName);
             }
             availableBdtPriRestri.setDefault(isDefault);
-            availableBdtPriRestri.setXbtId(e.get(XBT.XBT_ID).toBigInteger());
+            availableBdtPriRestri.setXbtId(e.get(XBT.XBT_ID));
             availableBdtPriRestri.setXbtName(xbtName);
             return availableBdtPriRestri;
         };
